Added new guard against casting errors in wing section definition

+++ LovelacePM/wing.py
@@ -20,12 +20,12 @@ class wing_section: #class to define wing section based on airfoil info
             correction=None, Re=2e6, sweep=0.0):
         #closed: return camberline as set of points. for closing wings
         #control axpercs: axis percentage of chord position
-        self.CA_position=CA_position
-        self.c=c
+        self.CA_position=CA_position.astype(float)
+        self.c=float(c)
         self.closed=closed
         self.points=wing_afl_positprocess(read_afl(afldir=afldir, afl=afl, ext_append=True, header_lines=header_lines, disc=xdisc, \
             remove_TE_gap=remove_TE_gap, incidence=radians(incidence), inverse=inverse, strategy=xstrategy, sweep=radians(sweep)), \
-                gamma=radians(gamma), c=c, xpos=CA_position[0], ypos=CA_position[1], zpos=CA_position[2])
+                gamma=radians(gamma), c=c, xpos=float(CA_position[0]), ypos=float(CA_position[1]), zpos=float(CA_position[2]))
         self.inverted=inverse
         self.correction=correction
         self.xdisc=int(np.size(self.points, axis=0)/2)
